requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
numpy>=2.0.0
opencv-python>=4.10.0
opencv-python-headless>=4.10.0
pillow>=10.0.0
matplotlib>=3.9.0
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.40.0
ultralytics>=8.2.0
qiskit>=2.0.0
qiskit-aer>=0.15.0
pylatexenc>=2.10
scipy>=1.13.0
streamlit>=1.35.0
python-multipart>=0.0.9
